pyproject.toml
[project]
name = "github-mcp-server"
version = "0.1.0"
description = "MCP server for GitHub API operations - issues, PRs, CI, milestones"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Rafael Riesco", email = "rriesco@users.noreply.github.com"}
]
keywords = ["mcp", "github", "claude", "anthropic", "api", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=0.9.0",
    "PyGithub>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/rriesco/github-mcp-server"
Repository = "https://github.com/rriesco/github-mcp-server"
Issues = "https://github.com/rriesco/github-mcp-server/issues"

[project.scripts]
github-mcp-server = "github_mcp_server.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/github_mcp_server"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
markers = [
    "unit: Fast unit tests with mocks (run on every commit)",
    "integration: Slower tests with real API calls (run nightly or before release)",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
